"Update payment controller, transaction routes, and demo role seeder"

+++ seeders/20240523145340-demo-role.js
@@ -3,15 +3,23 @@ const { v4: uuidv4 } = require('uuid');
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkInsert('roles', [{
-      id: uuidv4(),
-      name: 'user',
-      createdAt: new Date(),
-      updatedAt: new Date()
-    }], {});
+    await queryInterface.bulkInsert('roles', [
+      {
+        id: uuidv4(),
+        name: 'user',
+        createdAt: new Date(),
+        updatedAt: new Date()
+      },
+      {
+        id: uuidv4(),
+        name: 'admin',
+        createdAt: new Date(),
+        updatedAt: new Date()
+      }
+    ], {});
   },
 
   down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete('roles', { name: 'user' }, {});
+    await queryInterface.bulkDelete('roles', { name: ['user', 'admin'] }, {});
   }
 };
